Todo DetailView wurde erstellt.
