+++ source/public/scripts/views/TodoView.js
@@ -19,51 +19,10 @@ class TodoView {
 
         this.app.addEventListener('click', (event) => {
             if (event.target.id === 'create-todo-button') {
-                this.showTodoForm();
+                const createEvent = new CustomEvent('showCreateTodoForm');
+                document.dispatchEvent(createEvent);
             }
         });
-    }
-
-    showTodoForm() {
-        this.app.innerHTML = `
-            <form id="todo-form">
-                <label>Title:</label>
-                <input type="text" placeholder="Example">
-                <label>Importance:</label>
-                <input type="number" min="1" max="5" placeholder="5">
-                <label>Due Date:</label>
-                <input type="date">
-                <label>Completed:</label>
-                <input type="checkbox">
-                <label>Description:</label>
-                <textarea placeholder="Example Todo"></textarea>
-                <button type="submit">Create</button>
-            </form>
-        `;
-
-        document.getElementById('todo-form').addEventListener('submit', (event) => {
-            event.preventDefault();
-            this.handleFormSubmit();
-        });
-    }
-
-    handleFormSubmit() {
-        const title = document.querySelector('input[type="text"]').value;
-        const importance = document.querySelector('input[type="number"]').value;
-        const dueDate = document.querySelector('input[type="date"]').value;
-        const description = document.querySelector('textarea').value;
-        const completed = document.querySelector('input[type="checkbox"]').checked;
-
-        const newTodo = {
-            title,
-            importance,
-            dueDate,
-            description,
-            completed
-        };
-
-        const event = new CustomEvent('todoCreated', { detail: newTodo });
-        document.dispatchEvent(event);
     }
 
     createTodoList(todos) {
@@ -97,37 +56,6 @@ class TodoView {
                 const eventDetail = new CustomEvent('editTodo', { detail: { todo: todos[index], index: parseInt(index, 10) } });
                 document.dispatchEvent(eventDetail);
             }
-        });
-    }
-
-    showEditForm(todo) {
-        this.app.innerHTML = `
-            <form id="todo-form">
-                <label>Title:</label>
-                <input type="text" value="${todo.title}">
-                <label>Importance:</label>
-                <input type="number" min="1" max="5" value="${todo.importance}">
-                <label>Due Date:</label>
-                <input type="date" value="${todo.dueDate}">
-                <label>Completed:</label>
-                <input type="checkbox" ${todo.completed ? 'checked' : ''}>
-                <label>Description:</label>
-                <textarea>${todo.description}</textarea>
-                <button type="submit">Update</button>
-            </form>
-        `;
-
-        document.getElementById('todo-form').addEventListener('submit', (event) => {
-            event.preventDefault();
-            const updatedTodo = {
-                title: document.querySelector('input[type="text"]').value,
-                importance: document.querySelector('input[type="number"]').value,
-                dueDate: document.querySelector('input[type="date"]').value,
-                description: document.querySelector('textarea').value,
-                completed: document.querySelector('input[type="checkbox"]').checked
-            };
-            const updateEvent = new CustomEvent('updateTodo', { detail: { updatedTodo, index: todo.index } });
-            document.dispatchEvent(updateEvent);
         });
     }
 }
